+++ packages/connect/src/actions/app/index.ts
@@ -1,4 +1,0 @@
-export * from "./connect";
-export * from "./status";
-export * from "./verifySignInMessage";
-export * from "./watchStatus";